refactor: break down blockdetail into smaller components
